new normalization final adjustments
