+++ lib/meta.js
@@ -2,15 +2,19 @@
  * Meta Extension
  *
  * Sets default metadata on a load record (load.metadata) from
- * loader.meta.
+ * loader.metadata via System.meta function.
  *
  *
  * Also provides an inline meta syntax for module meta in source.
  *
  * Eg:
  *
- * loader.meta['my/module'] = { deps: ['jquery'] };
- * loader.meta['my/*'] = { format: 'amd' };
+ * loader.meta({
+ *   'my/module': { deps: ['jquery'] }
+ *   'my/*': { format: 'amd' }
+ * });
+ * 
+ * Which in turn populates loader.metadata.
  *
  * load.metadata.deps and load.metadata.format will then be set
  * for 'my/module'
@@ -26,6 +30,14 @@
  * Configuration meta always takes preference to inline meta.
  *
  * Multiple matches in wildcards are supported and ammend the meta.
+ *
+ *
+ * The benefits of the function form is that paths are URL-normalized
+ * supporting say
+ *
+ * loader.meta({ './app': { format: 'cjs' } });
+ *
+ * Instead of needing to set against the absolute URL (https://site.com/app.js)
  * 
  */
 
@@ -34,16 +46,19 @@
   hookConstructor(function(constructor) {
     return function() {
       constructor.call(this);
-      this.meta = {};
+      this.metadata = {};
     };
   });
 
-  function extend(a, b, overwrite) {
-    for (var p in b) {
-      if (overwrite || !(p in a))
-        a[p] = b[p];
+  SystemJSLoader.prototype.meta = function(metadata) {
+    for (var module in metadata) {
+      var meta = metadata[module];
+      // normalize relative URL module specifiers to baseURL
+      if (module[0] == '/' || module[0] == '.')
+        module = new URL(module, getAbsBaseURL.call(this)).href;
+      this.metadata[module] = meta;
     }
-  }
+  };
 
   hook('locate', function(locate) {
     return function(load) {
@@ -65,13 +80,13 @@
           var depth = module.split('/').length;
           if (depth > bestDepth)
             bestDetph = depth;
-          extend(load.metadata, meta[module], bestDepth == depth);
+          extend(load.metadata, meta[module], bestDepth != depth);
         }
       }
 
       // apply exact meta
       if (meta[name])
-        extend(load.metadata, meta[name], true);
+        extend(load.metadata, meta[name]);
 
       return locate.call(this, load);
     };
