+++ lib/package.js
@@ -7,29 +7,28 @@
  *
  * Example:
  *
- * System.config({
- *   packages: {
- *     jquery: {
- *       main: 'index.js', // this main is actually set by default
- *       format: 'amd',
- *       defaultExtension: 'js',
- *       meta: {
- *         '*.ts': {
- *           plugin: 'typescript'
- *         },
- *         'vendor/sizzle.js': {
- *           format: 'global'
- *         }
+ * System.package({
+ *   jquery: {
+ *     main: 'index.js', // this main is actually set by default
+ *     format: 'amd',
+ *     defaultExtension: 'js',
+ *     meta: {
+ *       '*.ts': {
+ *         plugin: 'typescript'
  *       },
- *       map: {
- *         // map internal require('sizzle') to local require('./vendor/sizzle')
- *         sizzle: './vendor/sizzle.js',
- *
- *         // map any internal or external require of 'jquery/vendor/another' to 'another/index.js'
- *         './vendor/another': 'another/index.js'
+ *       'vendor/sizzle.js': {
+ *         format: 'global'
  *       }
- *     }
- *   }
+ *     },
+ *     map: {
+ *        // map internal require('sizzle') to local require('./vendor/sizzle')
+ *        sizzle: './vendor/sizzle.js',
+ *
+ *        // map any internal or external require of 'jquery/vendor/another' to 'another/index.js'
+ *        './vendor/another': 'another/index.js'
+ *      }
+ *    }
+ *  }
  * });
  *
  * Then:
@@ -52,6 +51,10 @@
       this.packages = {};
     };
   });
+
+  SystemJSLoader.prototype.package = function(packages) {
+    extend(this.packages, packages);
+  };
 
   function getPackage(name) {
     for (var p in this.packages) {
@@ -131,13 +134,6 @@
     };
   });
 
-  function extend(a, b, overwrite) {
-    for (var p in b) {
-      if (overwrite || !(p in a))
-        a[p] = b[p];
-    }
-  }
-
   hook('locate', function(locate) {
     return function(load) {
       var loader = this;
@@ -165,14 +161,14 @@
                 var depth = module.split('/').length;
                 if (depth > bestDepth)
                   bestDetph = depth;
-                extend(meta, pkg.meta[module], bestDepth == depth);
+                extend(meta, pkg.meta[module], bestDepth != depth);
               }
             }
             // exact meta
             if (pkg.meta[load.name])
-              extend(meta, meta[load.name], true);
+              extend(meta, meta[load.name]);
             
-            extend(load.metadata, meta);
+            extend(load.metadata, meta, true);
           }
         }
