+++ lib/plugins.js
@@ -85,7 +85,7 @@
             metadata: {}
           };
           for (var p in load.metadata) {
-            if (p !== 'plugin')
+            if (p !== 'loader')
             argLoad.metadata[p] = load.metadata[p];
           }
           if (pluginModule.locate)
