+++ lib/proto.js
@@ -1,0 +1,36 @@
+// SystemJS Loader Class and Extension helpers
+
+function SystemJSLoader(options) {
+  SystemLoader.call(this, options);
+
+  systemJSConstructor.call(this, options);
+}
+
+// inline Object.create-style class extension
+function SystemProto() {};
+SystemProto.prototype = SystemLoader.prototype;
+SystemJSLoader.prototype = new SystemProto();
+
+var systemJSConstructor;
+
+function hook(name, hook) {
+  SystemJSLoader.prototype[name] = hook(SystemJSLoader.prototype[name]);
+}
+function hookConstructor(hook) {
+  systemJSConstructor = hook(systemJSConstructor || function() {});
+}
+
+function dedupe(deps) {
+  var newDeps = [];
+  for (var i = 0, l = deps.length; i < l; i++)
+    if (indexOf.call(newDeps, deps[i]) == -1)
+      newDeps.push(deps[i])
+  return newDeps;
+}
+
+function extend(a, b, underwrite) {
+  for (var p in b) {
+    if (!underwrite || !(p in a))
+      a[p] = b[p];
+  }
+}
